+++ src/DimensionReader/DimensionsReader.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Guillaumetissier\ImageResizer\DimensionReader;
 
+use Guillaumetissier\ImageResizer\Constants\ImageType;
 use Guillaumetissier\ImageResizer\Exceptions\CannotReadImageSizeException;
+use Guillaumetissier\ImageResizer\Exceptions\InvalidImageTypeException;
 use Guillaumetissier\ImageResizer\Exceptions\InvalidPathException;
 use Guillaumetissier\ImageResizer\ImageDimensions;
 use Guillaumetissier\PathUtilities\Path;
@@ -13,6 +15,7 @@ final class DimensionsReader implements DimensionsReaderInterface
 {
     /**
      * @throws InvalidPathException|CannotReadImageSizeException
+     * @throws InvalidImageTypeException
      */
     public function readDimensions(Path $file): ImageDimensions
     {
@@ -22,6 +25,10 @@ final class DimensionsReader implements DimensionsReaderInterface
 
         if (false === ($infos = @getimagesize((string) $file))) {
             throw new CannotReadImageSizeException($file);
+        }
+
+        if (!in_array($infos[2], ImageType::all(), true)) {
+            throw InvalidImageTypeException::invalidImageType(ImageType::toExtension($infos[2]), ImageType::allExtensions());
         }
 
         return new ImageDimensions($infos[1], $infos[0]);