+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2026-01-27
+## [1.1.1] - 2026-01-30
+
+### Fixed
+- Add missing image type validation in DimensionsReader
+- Prevent processing of unsupported image formats
+
+### Added
+- InvalidImageTypeException for clearer error messages when using unsupported formats
+
+## [1.1.0] - 2026-01-30
 
 ### Added